requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlmodel==0.0.22
python-dotenv==1.0.1
scikit-learn==1.6.1
python-multipart==0.0.20
python-pptx==1.0.2
python-docx==1.2.0
joblib==1.4.2
